AgentePessoal/pyproject.toml
[project]
name = "agent-skill-python"
version = "1.0.0"
description = "Agente de Habilidades ReAct Multi-Modal para Telegram"
authors = [
    { name = "Wesley Faria", email = "wesley.faria@detran.go.gov.br" }
]
dependencies = [
    "python-telegram-bot>=21.0",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1",
    "pypdf>=4.0.0",
    "httpx>=0.26.0",
    "duckduckgo-search>=4.4.2",
    "beautifulsoup4>=4.12.3",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
